+++ src/layouts/explorer/workflow/index.js
@@ -512,28 +512,6 @@ function WorkingRevision(props) {
                         </FlexBox>
                     </FlexBox>:""}
                     {tabBtn === 1 ? <DiagramEditor workflow={oldWf} namespace={namespace} updateWorkflow={(data)=>{
-                        // const workflowStr = PrettyYAML.stringify(data)
-                        // let newWorkflowStr = ""
-                        
-
-                        // workflowStr.split("\n").forEach(lineStr => {
-                        //     let whiteSpaceCount = 0
-                        //     for (;  lineStr[whiteSpaceCount] === " "; whiteSpaceCount++) {}
-                        //     let processedLineStr = lineStr.slice(whiteSpaceCount).trim().slice(0, -1)
-                        //     console.log("processedLineStr = ", processedLineStr)
-                        //     if (processedLineStr.startsWith(`transform:`)) {
-                        //         const jsStr = processedLineStr.slice(whiteSpaceCount+`transform: "jq(`.length)
-                        //         newWorkflowStr += " ".repeat(whiteSpaceCount) + "transform: |\n"
-                        //         newWorkflowStr += " ".repeat(whiteSpaceCount) + "  js(" + "\n"
-                        //         jsStr.split("\\n").forEach(jsLineStr => {
-                        //             newWorkflowStr += " ".repeat(whiteSpaceCount) + "    " + jsLineStr + "\n"
-                        //         });
-                        //         newWorkflowStr += " ".repeat(whiteSpaceCount) + "  )" + "\n"
-                        //     } else {
-                        //         newWorkflowStr += lineStr + "\n"
-                        //     }
-                        // });
-
                         setWorkflow(data)
 
                         setTabBtn(0)
